split concat tokenization into data
